Cleanup upgrade_manifest_json code
